Release gui mutex before clipboard keep-alive wait
The COSMIC/GNOME clipboard workaround added in #4902 keeps the
capture window alive (hidden) for up to 30s while waiting for a
paste consumer to fetch the clipboard data. That process still held
the 'flameshot gui' single-instance lock (QSharedMemory-based, see
guiMutexLock() in main.cpp) for the entire wait, so a second
'flameshot gui' invocation during that window silently failed to
acquire the lock and exited with no visible UI at all.

Moves guiMutexLock() into its own small utils/guimutex.{h,cpp} and
adds releasePendingGuiMutex(), an idempotent release callable from
both the normal aboutToQuit path and, now, from the clipboard
workaround itself: the mutex is released as soon as we hand off to
the keep-alive wait, since the actual GUI portion of the capture is
already done at that point and a new capture shouldn't be blocked by
one that's only lingering to serve a pending clipboard read.

Credit to @cleytonmendest for finding and precisely diagnosing this
in review on #4902.

Verified on Pop!_OS 24.04 with COSMIC 1.0.0: rapid back-to-back
Ctrl+Shift+S captures now both work immediately, where previously
the second one silently did nothing until the first's 30s timeout
(or an incidental clipboard-ownership change) released the lock.
